Add some more documentation
