+++ listenbrainz/webserver/views/views_utils.py
@@ -3,6 +3,9 @@ from flask_login import current_user
 
 
 def get_current_spotify_user():
+    """Returns the spotify access token and permissions for the current
+    authenticated user. If the user is not authenticated or has not linked
+    to a Spotify account, returns empty dict."""
     if not current_user.is_authenticated:
         return {}
     user = SpotifyService().get_user(current_user.id)
